% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_kendall_fit.R
\name{addMultiKendall}
\alias{addMultiKendall}
\title{Fit stable Kendall distribution to multiple sets of observations}
\usage{
addMultiKendall(srcFrame, alpha, normalise = FALSE,
  groupingVariables = c("year", "polutant"), m_alpha = function(x) x,
  symmetric = FALSE)
}
\arguments{
\item{srcFrame}{tibble returned by calculateMaxima function}

\item{alpha}{value of alpha parameter}

\item{normalise}{If TRUE, maximums will be scaled and centered.}

\item{groupingVariables}{chr, vector of names of columns to group by,
year and polutant by default}

\item{m_alpha}{function giving moment of order alpha of the step distribution}

\item{symmetric}{if TRUE, symmetrical version of stable Kendall distribution will be used}
}
\value{
tibble with empirical CDF, theoretical CDF and theoretical quantiles.
}
\description{
Fit stable Kendall distribution to multiple sets of observations
}
