% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxima.R
\name{calculateMaxima}
\alias{calculateMaxima}
\title{Calculate maxima for a given station and time step.}
\usage{
calculateMaxima(sourceFrame, block = "day")
}
\arguments{
\item{sourceFrame}{tibble returned by importGiosFromXLSX function.}

\item{block}{time step for maxima calculations.}
}
\value{
tibble with maxima grouped by year, date, polutant.
}
\description{
Calculate maxima for a given station and time step.
}
\examples{
obs <- importGiosFromXLSX("DsWrocKorzA", c("NOx", "SO2"), c("2015", "2014"))
obs2 <- calculateMaxima(obs)

}
