% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{cdfsGev}
\alias{cdfsGev}
\title{Compare theorical and empirical CDF for fitted GEV distribution.}
\usage{
cdfsGev(sourceFrame, fittedGEV)
}
\arguments{
\item{sourceFrame}{tibble returned by filteredData()}

\item{fittedGEV}{list returned by egevd function}
}
\value{
ggplot2 object
}
\description{
Compare theorical and empirical CDF for fitted GEV distribution.
}
