% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{cdfsKendall}
\alias{cdfsKendall}
\title{Compare theorical and empirical CDF for stable Kendall distribution.}
\usage{
cdfsKendall(sourceFrame)
}
\arguments{
\item{sourceFrame}{tibble returned by addMultiKendall()}
}
\value{
ggplot2 object
}
\description{
Compare theorical and empirical CDF for stable Kendall distribution.
}

