% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{convergenceVis}
\alias{convergenceVis}
\title{Draw simulated trajectories}
\usage{
convergenceVis(simulations, ogrX = NULL)
}
\arguments{
\item{simulations}{tibble returned by normingSequences function}

\item{ogrX}{maximum value on x axis}
}
\value{
ggplot2 object
}
\description{
Draw simulated trajectories
}

