% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendDist.R
\name{dkend}
\alias{dkend}
\title{PDF of Kendall stable distribution}
\usage{
dkend(mAlpha)
}
\arguments{
\item{mAlpha}{function giving moments of order alpha of step dist.}
}
\value{
function that returns values of the PDF
}
\description{
PDF of Kendall stable distribution
}

