% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_separately.R
\name{estimate_location}
\alias{estimate_location}
\title{Estimate location parameter of generalized stable Kendall distribution}
\usage{
estimate_location(data, m_alpha)
}
\arguments{
\item{data}{numeric vector of observations}

\item{m_alpha}{m_alpha function to be used}
}
\value{
estimated value of location parameter
}
\description{
Estimate location parameter of generalized stable Kendall distribution
}
