% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_separately.R
\name{estimate_scale}
\alias{estimate_scale}
\title{Estimate scale parameter of generalized stable Kendall distribution}
\usage{
estimate_scale(data, m_alpha)
}
\arguments{
\item{data}{numeric vector of observations}

\item{m_alpha}{m_alpha function to be used}
}
\value{
estimated value of scale parameter
}
\description{
Estimate scale parameter of generalized stable Kendall distribution
}
