% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_kendall_fit.R
\name{estimate_stable_alpha}
\alias{estimate_stable_alpha}
\title{Fit alpha parameter using MLE for distribution with one parameter}
\usage{
estimate_stable_alpha(data)
}
\arguments{
\item{data}{Numeric vector}
}
\value{
list with optimal parameter and loglikelihood value
}
\description{
Fit alpha parameter using MLE for distribution with one parameter
}
