% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_kendall_fit.R
\name{fit_stable_alpha}
\alias{fit_stable_alpha}
\title{Fit stable Kendall distribution with one parameter (alpha)}
\usage{
fit_stable_alpha(data)
}
\arguments{
\item{data}{Numeric vector of data.}
}
\value{
list of type kendall_fit
}
\description{
Fit stable Kendall distribution with one parameter (alpha)
}
