% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitMultiGEV}
\alias{fitMultiGEV}
\title{Fit GEV distribution to multiple sets of observations}
\usage{
fitMultiGEV(srcFrame, groupingVariables = c("year", "polutant"),
  full = TRUE)
}
\arguments{
\item{srcFrame}{tibble returned by calculateMaxima function}

\item{groupingVariables}{chr, vector of names of columns to group by,
year and polutant by default}

\item{full}{lgl, if TRUE, tibble from srcFrame argument will be returned
with additional columns giving parameter values of fitted GEV dist.
If FALSE, just parameters (with corresponding values of grouping variables)
will be returned.}

\item{...}{additional parameters passed to egevd() function}
}
\value{
tibble as described in help for full argument
}
\description{
Fit GEV distribution to multiple sets of observations
}

