% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_kendall_fit.R
\name{full_loglik_gradient}
\alias{full_loglik_gradient}
\title{Gradient of minus loglikelihood for stable Kendall distribution with 3 parameters.}
\usage{
full_loglik_gradient(data)
}
\arguments{
\item{data}{numeric vector of observation.}
}
\value{
Function of one argument of length 3 (alpha, location, scale).
}
\description{
Gradient of minus loglikelihood for stable Kendall distribution with 3 parameters.
}
