% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_separately.R
\name{full_minus_loglik}
\alias{full_minus_loglik}
\title{Negative loglikelihood for stable Kendall distr. with 3 parameters.}
\usage{
full_minus_loglik(data, parameters)
}
\arguments{
\item{parameters}{}
}
\value{
numeric, value of loglikelihood
}
\description{
Negative loglikelihood for stable Kendall distr. with 3 parameters.
}
