% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_ladder_moments.R
\name{g_function}
\alias{g_function}
\title{Function G(t) - Williamson transform taken at point 1/t.}
\usage{
g_function(t, alpha, density)
}
\arguments{
\item{t}{Argument to the function.}

\item{alpha}{Value of the alpha parameter.}

\item{density}{Density function of the step distribution.}
}
\value{
Object of class "integrate"
}
\description{
Function G(t) - Williamson transform taken at point 1/t.
}
