% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_ladder_moments.R
\name{g_function_single}
\alias{g_function_single}
\title{Function G(t) - Williamson transform taken at point 1/t.}
\usage{
g_function_single(t, alpha, density)
}
\arguments{
\item{t}{Argument to the function.}

\item{alpha}{Value of the alpha parameter.}

\item{density}{Density function of the step distribution.}
}
\value{
Object of class "integrate"
}
\description{
This function return the whole "integrate" object, so precision of the approximation
can be checked.
}
\examples{
g_function_single(5, 0.26, dnorm)

}
