% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importGiosFromCSV}
\alias{importGiosFromCSV}
\title{Import data for one station, multiple years and polutants from .csv files.}
\usage{
importGiosFromCSV(station, polutants = NULL, years = NULL, path = getwd(),
  noHours = "1", skip = 3, exact = FALSE)
}
\arguments{
\item{station}{chr, name of the chosen station.}

\item{polutants}{chr, names of chosen polutants.}

\item{years}{chr, chosen years (yyyy).}

\item{path}{chr, path to folder containing .xlsx file or path to .xlsx file.}

\item{noHours}{chr, 1 or 24, defaults to 1 (hourly meas.)}

\item{skip}{int, number of rows to skip, defaults to 3.
Do not change if unchanged GIOŚ files are used.}

\item{exact}{lgl, TRUE if path is a path to .xlsx file.}
}
\value{
tibble
}
\description{
Import data for one station, multiple years and polutants from .csv files.
}
\examples{
importGiosFromXLSX("DsWrocKorzA", c("NOx", "SO2"), c("2015", "2014"))
# Default settings, .xlsx files are in the working directory.
importGiosFromXLSX("DsWrocKorzA", c("NOx", "SO2"), c("2015", "2014"), "path-to-the-folder")
# Importing from a different (not working) directory.

}

