% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importOneCSV}
\alias{importOneCSV}
\title{Import from one .csv file.}
\usage{
importOneCSV(station, polutant, year, path = getwd(), noHours = "1",
  skip = 3, exact = FALSE)
}
\arguments{
\item{station}{chr, name of the chosen station.}

\item{polutant}{chr, name of chosen polutant.}

\item{year}{chr, chosen year (yyyy).}

\item{path}{chr, path to folder containing .xlsx file or path to .xlsx file.}

\item{noHours}{chr, 1 or 24, defaults to 1 (hourly meas.)}

\item{skip}{int, number of rows to skip, defaults to 3.
Do not change if unchanged GIOŚ files are used.}

\item{exact}{lgl, TRUE if path is a path to .xlsx file.}
}
\value{
tibble
}
\description{
Import from one .csv file.
}

