% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importOneXLSX}
\alias{importOneXLSX}
\title{Import one file from GIOŚ .xlsx file.}
\usage{
importOneXLSX(station, polutant, year, noHours, path = getwd(), skip = 3,
  exact = FALSE)
}
\arguments{
\item{station}{chr, name of the chosen station.}

\item{polutant}{chr, name of chosen polutant.}

\item{year}{chr, chosen year (yyyy).}

\item{noHours}{chr, 1 or 24, defaults to 1 (hourly meas.)}

\item{path}{chr, path to folder containing .xlsx file or path to .xlsx file.}

\item{skip}{int, number of rows to skip, defaults to 3.
Do not change if unchanged GIOŚ files are used.}

\item{exact}{lgl, TRUE if path is a path to .xlsx file.}
}
\value{
tibble
}
\description{
Import one file from GIOŚ .xlsx file.
}

