% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{isAvailable}
\alias{isAvailable}
\title{Check if there are any measurement on a given stations for chosen year and polutant.}
\usage{
isAvailable(chosenYear, chosenPolutant, station)
}
\arguments{
\item{chosenYear}{year of measurements}

\item{chosenPolutant}{measured polutant}

\item{station}{station of interest. Please provide current names of stations (values, not names of stationCodes vector)}
}
\value{
logical value: TRUE if the data is available, FALSE if it's not
}
\description{
Check if there are any measurement on a given stations for chosen year and polutant.
}

