% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_separately.R
\name{kendall_loglik}
\alias{kendall_loglik}
\title{Log-likelihood for stable kendall distribution with m_alpha = 1}
\usage{
kendall_loglik(alpha, x)
}
\arguments{
\item{alpha}{alpha parameter of the Kendall random walk}

\item{x}{numeric vector of observations}
}
\value{
numeric
}
\description{
Log-likelihood for stable kendall distribution with m_alpha = 1
}
