% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendallRandomPackage.R
\docType{package}
\name{kendallRandomPackage}
\alias{kendallRandomPackage}
\alias{kendallRandomPackage-package}
\title{kendallRandomPackage: package for exploring air pollution data in Poland.}
\description{
Kendall random walks are Markov processes generated by the Kendall convolution.
This package helps simulate and visualize these random walks and associated
distributions. It also provides function to fit these distribution to data.
}
