% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_ladder_moments.R
\name{ladder_moment_pmf}
\alias{ladder_moment_pmf}
\title{Distribution of the first ladder moment.}
\usage{
ladder_moment_pmf(n, level, alpha, step_cdf, step_pdf)
}
\arguments{
\item{n}{Argument to the PDF.}

\item{level}{Level a to be crossed.}

\item{alpha}{Alpha parameter of Kendall random walk.}

\item{step_cdf}{CDF of the step distribution.}

\item{step_pdf}{PDF of the step distribution.}
}
\value{
Value of PMF of the distribution of first ladder moment
}
\description{
Distribution of the first ladder moment.
}
\examples{
prob <- ladder_moment_pmf(10, 1000, 0.5, pnorm, dnorm)
prob


}
