% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations_kendall_rw.R
\name{mutate_kendall_rw}
\alias{mutate_kendall_rw}
\title{Mutate each trajectory.}
\usage{
mutate_kendall_rw(simulations, mutate_function, df = T)
}
\arguments{
\item{simulations}{Object of class kendall_simulation.}

\item{mutate_function}{Function that will be applied to each trajectory.}

\item{df}{If TRUE, a d.f will be returned, if FALSE, simulations in the kendall_simulation
object passed in simulations argument will be replaced by the result of mutate_function.}
}
\value{
data frame or a list (of class kendall_simulation)
}
\description{
Mutate each trajectory.
}
