% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{normalising_sequences}
\alias{normalising_sequences}
\title{Normalising Kendall random walks}
\usage{
normalising_sequences(simulations, an_seq = 1, bn_seq = 0)
}
\arguments{
\item{simulations}{tibble returned by simulation function}

\item{an_seq}{sequence that the trajectories will be multiplied by}

\item{bn_seq}{sequence that will be substracted from scaled trajectory}
}
\value{
tibble
}
\description{
Normalising Kendall random walks
}
