% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{normingSequences}
\alias{normingSequences}
\title{Normalising Kendall random walks}
\usage{
normingSequences(simulations, AnSeq = 1, BnSeq = 0)
}
\arguments{
\item{simulations}{tibble returned by simulation function}

\item{AnSeq}{sequence that the trajectories will be multiplied by}

\item{BnSeq}{sequence that will be substracted from scaled trajectory}
}
\value{
tibble
}
\description{
Normalising Kendall random walks
}

