% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendDist.R
\name{pkend}
\alias{pkend}
\title{CDF of Kendall stable distribution}
\usage{
pkend(mAlpha)
}
\arguments{
\item{mAlpha}{function giving moments of order alpha of step dist.}
}
\value{
function function giving values of CDF of Kendall stable distribution
}
\description{
CDF of Kendall stable distribution
}

