% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier_crossing.R
\name{plot.kendall_barrier_crossing}
\alias{plot.kendall_barrier_crossing}
\title{Generic function for plotting results of ladder_moment function.}
\usage{
\method{plot}{kendall_barrier_crossing}(x, ...)
}
\arguments{
\item{x}{kendall_barrier_crossing object}

\item{...}{Additional arguments}
}
\value{
ggplot2
}
\description{
Generic function for plotting results of ladder_moment function.
}
