% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_kendall_fit.R
\name{plot.kendall_fit}
\alias{plot.kendall_fit}
\title{QQ-plot for the result of fitting stable Kendall distribtion.}
\usage{
\method{plot}{kendall_fit}(x, ...)
}
\arguments{
\item{x}{List returned by fit_separate or fit_kendall function.}

\item{...}{Aditional arguments.}
}
\value{
ggplot2 object
}
\description{
QQ-plot for the result of fitting stable Kendall distribtion.
}
