% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{plot.kendall_simulation}
\alias{plot.kendall_simulation}
\title{Draw simulated trajectories}
\usage{
\method{plot}{kendall_simulation}(x, max_x = NULL, level = NULL, ...)
}
\arguments{
\item{x}{object returned by normalising_sequences function.}

\item{max_x}{maximum value on x axis.}

\item{level}{Y-axis value which will be marked (level to be crossed).}

\item{...}{Other arguments}
}
\value{
ggplot2 object
}
\description{
Draw simulated trajectories
}
