% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations_kendall_rw.R
\name{plot.kendall_summary}
\alias{plot.kendall_summary}
\title{Plot summary of Kendall random walk simulations.}
\usage{
\method{plot}{kendall_summary}(x, ..., type = "density")
}
\arguments{
\item{x}{Object of class kendall_summary}

\item{...}{Optional arguments, currently ignored}

\item{type}{Type of the plot: density, histogram or boxplot}
}
\value{
ggplot2 object
}
\description{
Plot summary of Kendall random walk simulations.
}
