% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plotHist}
\alias{plotHist}
\title{Plot histogram of values over given threshold}
\usage{
plotHist(srcTbl, threshold)
}
\arguments{
\item{srcTbl}{tibble returned by filteredData() interactive}

\item{threshold}{cut-off value}
}
\value{
ggplot2 object
}
\description{
Plot histogram of values over given threshold
}

