% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plotLargeQQ}
\alias{plotLargeQQ}
\title{QQ-plot for large quantiles}
\usage{
plotLargeQQ(srcTbl, alpha, minMaxQ, stepQ, meanFunction = function(x) 1)
}
\arguments{
\item{srcTbl}{tibble returned by filteredData()}

\item{alpha}{Kendall stable distribution parameter}

\item{minMaxQ}{minimum and maximum quantile to be used}

\item{stepQ}{step between minimum and maximum quantile}

\item{meanFunction}{function giving moment of order alpha of the step distribution}
}
\value{
ggplot2 object
}
\description{
QQ-plot for large quantiles
}

