% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plotTime}
\alias{plotTime}
\title{Plot of data over time}
\usage{
plotTime(srcTbl, datesRange = "")
}
\arguments{
\item{srcTbl}{tibble returned by filteredData()}
}
\value{
ggplot2 object
}
\description{
Plot of data over time
}
