% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier_crossing.R
\name{print.kendall_barier_crossing}
\alias{print.kendall_barier_crossing}
\title{Generic function for printing result of ladder_moment function}
\usage{
\method{print}{kendall_barier_crossing}(x, ...)
}
\arguments{
\item{x}{kendall_barrier_crossing object}

\item{...}{Additional arguments}
}
\value{
invisible x
}
\description{
Generic function for printing result of ladder_moment function
}
