% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations_kendall_rw.R
\name{print.kendall_simulation}
\alias{print.kendall_simulation}
\title{Generic function that prints information about simulated Kendall random walk}
\usage{
\method{print}{kendall_simulation}(x, ...)
}
\arguments{
\item{x}{Object returned by simulate_kendall_rw or transform_kendall_rw function.}

\item{...}{Other arguments.}
}
\description{
Generic function that prints information about simulated Kendall random walk
}
