% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations_kendall_rw.R
\name{print.kendall_summary}
\alias{print.kendall_summary}
\title{Print summary of Kendall random walk simulations.}
\usage{
\method{print}{kendall_summary}(x, ...)
}
\arguments{
\item{x}{Object of type kendall_summary}

\item{...}{Optional parameters, currently ignored}
}
\description{
Print summary of Kendall random walk simulations.
}
