% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendDist.R
\name{qkend}
\alias{qkend}
\title{Quantiles of Kendall stable distribution}
\usage{
qkend(mAlpha)
}
\arguments{
\item{mAlpha}{function giving moments of order alpha of step dist.}
}
\value{
function function returning quantiles of given orders
}
\description{
Quantiles of Kendall stable distribution
}

