% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendDist.R
\name{qkendSym}
\alias{qkendSym}
\title{Quantiles of symmetrical Kendall stable distribution}
\usage{
qkendSym(mAlpha)
}
\arguments{
\item{mAlpha}{function giving moments of order alpha of step dist.}
}
\value{
function function returning quantiles of given orders
}
\description{
Quantiles of symmetrical Kendall stable distribution
}
\examples{
{
qKend <- qkendSym(function(x) 1)
# Step distribution: delta_{1}
qKendall <- qKend(c(0.1, 0.9), 0.5)
# Quantiles of order 0.1 and 0.9 for alpha = 0.5
}

}
