% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{qqPlotGev}
\alias{qqPlotGev}
\title{QQ-plot for fited GEV distribution}
\usage{
qqPlotGev(sourceFrame, fittedGEV)
}
\arguments{
\item{sourceFrame}{tibble returned by filteredData()}

\item{fittedGEV}{list returned by egevd function}
}
\value{
ggplot2 object
}
\description{
QQ-plot for fited GEV distribution
}
