% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendDist.R
\name{rkend}
\alias{rkend}
\title{Pseudo-random number from Kendall stable distribution}
\usage{
rkend(m_alpha)
}
\arguments{
\item{m_alpha}{function giving moments of order alpha of step dist.}
}
\value{
function return n numbers genereted from Kendall stable dist.
}
\description{
Pseudo-random number from Kendall stable distribution
}
