% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulation}
\alias{simulation}
\title{Simulate multiple trajectories of Kendall random walk}
\usage{
simulation(number_of_simulations, trajectory_length, step_dist, alpha, ...)
}
\arguments{
\item{number_of_simulations}{number of trajectories to generate.}

\item{trajectory_length}{length of trajectories.}

\item{step_dist}{function returning random numbers from step dist.}

\item{alpha}{alpha parameter.}

\item{...}{parameters for step distribution.}
}
\description{
Simulate multiple trajectories of Kendall random walk
}
