% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations_kendall_rw.R
\name{U}
\alias{U}
\title{Helper function}
\usage{
U(x, y)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}
}
\value{
sign of the argument whose abs. val. is bigger
}
\description{
Helper function
}
