% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{visualize_convergence}
\alias{visualize_convergence}
\title{Draw simulated trajectories}
\usage{
visualize_convergence(simulations, max_x = NULL)
}
\arguments{
\item{simulations}{tibble returned by normalising_sequences function}

\item{max_x}{maximum value on x axis}
}
\value{
ggplot2 object
}
\description{
Draw simulated trajectories
}
