% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{addMultiKendall}
\alias{addMultiKendall}
\title{Fit stable Kendall distribution to multiple sets of observations}
\usage{
addMultiKendall(srcFrame, groupingVariables = c("year", "polutant"),
  mAlpha = function(x) 1, parAlpha)
}
\arguments{
\item{srcFrame}{tibble returned by calculateMaxima function}

\item{groupingVariables}{chr, vector of names of columns to group by,
year and polutant by default}

\item{mAlpha}{function giving moment of order alpha of the step distribution}

\item{parAlpha}{value of alpha parameter}
}
\value{
tibble with empirical CDF, theoretical CDF and theoretical quantiles.
}
\description{
Fit stable Kendall distribution to multiple sets of observations
}

