% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendDist.R
\name{dkend}
\alias{dkend}
\title{PDF of Kendall stable distribution}
\usage{
dkend(m_alpha)
}
\arguments{
\item{m_alpha}{function giving moments of order alpha of step dist.}
}
\value{
function that returns values of the PDF
}
\description{
PDF of Kendall stable distribution
}
\examples{
{
dKend <- dkend(function(x) 1)
# Step distribution: delta_{1}
dKendall <- dKend(1:10, 0.5)
# Values of PDF for arguments 1:10 and alpha = 0.5
}

}
