% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_separately.R
\name{estimate_alpha}
\alias{estimate_alpha}
\title{Estimate alpha parameter of stable Kendall distribution}
\usage{
estimate_alpha(data, m_alpha, ...)
}
\arguments{
\item{data}{numeric vector of observations}

\item{m_alpha}{m_alpha function to be used}
}
\value{
estimated value of alpha parameter
}
\description{
Estimate alpha parameter of stable Kendall distribution
}
