% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_separately.R
\name{estimate_alpha}
\alias{estimate_alpha}
\title{Optimize log-likelihood to find alpha parameter}
\usage{
estimate_alpha(data)
}
\arguments{
\item{data}{numeric vector of observation}
}
\value{
numeric
}
\description{
Optimize log-likelihood to find alpha parameter
}
