% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_kendall_fit.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{Optimize log-likelihood to find distribution parameters}
\usage{
estimate_parameters(data, initial_point)
}
\arguments{
\item{data}{numeric vector of observation.}

\item{initial_point}{Initial guess for true parameters.}
}
\value{
Function of one argument of length 3 (alpha, location, scale)
        which will be used to solve nonlinear system of equations
        (find zeros of the derivative of log-likelihood).
}
\description{
Optimize log-likelihood to find distribution parameters
}
