% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_separately.R
\name{fit_kendall}
\alias{fit_kendall}
\title{Fit stable Kendall distribution for given data and m_alpha function.}
\usage{
fit_kendall(data, m_alpha, quantiles, initial_point)
}
\arguments{
\item{data}{Numeric vector of observation to which the distribution will be fitted.}

\item{m_alpha}{m_alpha function which will be used.}

\item{quantiles}{vector of quantile range to be estimated.}

\item{initial_point}{Initial guess for true parameters.}
}
\value{
fitted quantiles
}
\description{
Fit stable Kendall distribution for given data and m_alpha function.
}
