% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_kendall_fit.R
\name{fit_kendall}
\alias{fit_kendall}
\title{Fit stable Kendall distribution for given data and m_alpha function.}
\usage{
fit_kendall(data)
}
\arguments{
\item{data}{Numeric vector of observation to which the distribution will be fitted.}
}
\value{
fitted quantiles
}
\description{
Fit stable Kendall distribution for given data and m_alpha function.
}
