% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_kendall_fit.R
\name{fit_separate}
\alias{fit_separate}
\title{Function for fitting stable Kendall distribution separately to two parts of data}
\usage{
fit_separate(data, separation_point)
}
\arguments{
\item{data}{Numeric vector. Observation to which the distribution will be fitted.}

\item{separation_point}{Order above which data (quantiles) will be separated.}
}
\value{
List of class kendall_fit with estimated and theoretical quantiles
and estimated parameters.
}
\description{
Function for fitting stable Kendall distribution separately to two parts of data
}
