% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_kendall_fit.R
\name{full_minus_loglik}
\alias{full_minus_loglik}
\title{Negative loglikelihood for stable Kendall distr. with 3 parameters.}
\usage{
full_minus_loglik(data)
}
\arguments{
\item{data}{Dataset for which the loglikelihood will be calculated.}
}
\value{
numeric, value of loglikelihood
}
\description{
Negative loglikelihood for stable Kendall distr. with 3 parameters.
}
