% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{kendallRandomApp}
\alias{kendallRandomApp}
\title{Shiny app for exploring air pollution data}
\usage{
kendallRandomApp(sourceFrame)
}
\arguments{
\item{sourceFrame}{tibble returned by calculateMaxima function.}
}
\value{
shiny app object
}
\description{
Shiny app for exploring air pollution data
}

