% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{kendallRandomApp}
\alias{kendallRandomApp}
\title{Shiny app for exploring air pollution data}
\usage{
kendallRandomApp(sourceFrame, meanFunction = function(x) 1,
  symmetric = FALSE)
}
\arguments{
\item{sourceFrame}{tibble returned by calculateMaxima function.}

\item{meanFunction}{function giving moment of order alpha of Kendall stable distribution.}

\item{symmetric}{If TRUE, symmetrical version of Kendall stable distribution will be used.}
}
\value{
shiny app object
}
\description{
Shiny app for exploring air pollution data
}
\examples{
obs <- importGiosFromXLSX("DsWrocKorzA", c("NOx", "SO2"), c("2015", "2014"))
obs2 <- calculateMaxima(obs)
kendallRandomApp(obs2)

}
