% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_kendall_fit.R
\name{plot_qq}
\alias{plot_qq}
\title{Quick QQ-plot for the result of fitting two stable Kendall distributions.}
\usage{
plot_qq(fit_list)
}
\arguments{
\item{fit_list}{List returned by fit_separate function.}
}
\value{
ggplot2 object
}
\description{
Quick QQ-plot for the result of fitting two stable Kendall distributions.
}
