% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_kendall_distribution.R
\name{qkend}
\alias{qkend}
\title{Quantiles of Kendall stable distribution}
\usage{
qkend(m_alpha)
}
\arguments{
\item{m_alpha}{function giving moments of order alpha of step dist.}
}
\value{
function function returning quantiles of given orders
}
\description{
Quantiles of Kendall stable distribution
}
\examples{
qKend <- qkend(function(x) 1)
# Step distribution: delta_{1}
qKendall <- qKend(c(0.1, 0.9), 0.5)
# Quantiles of order 0.1 and 0.9 for alpha = 0.5


}
