% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulate_kendall_rw}
\alias{simulate_kendall_rw}
\title{Simulate multiple trajectories of Kendall random walk}
\usage{
simulate_kendall_rw(number_of_simulations, trajectory_length, step_dist, alpha,
  symmetric = FALSE, ...)
}
\arguments{
\item{number_of_simulations}{number of trajectories to generate.}

\item{trajectory_length}{length of trajectories.}

\item{step_dist}{function returning random numbers from step dist.}

\item{alpha}{alpha parameter.}

\item{symmetric}{If TRUE, random walk on the whole real line will be simulated.}

\item{...}{parameters for step distribution.}
}
\description{
Simulate multiple trajectories of Kendall random walk
}
