% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulateOneTrajectory}
\alias{simulateOneTrajectory}
\title{Simulate one trajectory of a Kendall random walk}
\usage{
simulateOneTrajectory(trajectoryLength, stepDist, parAlpha, ...)
}
\description{
Simulate one trajectory of a Kendall random walk
}

