% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulation}
\alias{simulation}
\title{Simulate multiple trajectories of Kendall random walk}
\usage{
simulation(simulationNumber, trajectoryLength, stepDist, parAlpha, ...)
}
\arguments{
\item{simulationNumber}{number of trajectories to generate}

\item{trajectoryLength}{length of trajectories}

\item{stepDist}{function returning random numbers from step dist.}

\item{parAlpha}{alpha parameter}
}
\description{
Simulate multiple trajectories of Kendall random walk
}

